Wall vegetation READ DESCRIPTION PLS
hey, the plants are in the game but as of right now some things still
need to resolved mainly regarding the actual sprites that in my opinion
could do with a bit more saturation on the stem and also darker outline.
Other than that i need to get some more of that sweet sweet variation in
there, mainly some moss, uma trepadeira, high plants and all kinds of
other stuff. I might however keep that for later bcause it is not the
most important priority and now that the system for it exists its
literaly a question of spending some time making the sprites. -I shall
now move on to jason combat and bug fixing :)